requirements.txt
aiosqlite==0.15.0
appdirs==1.4.4
apsw-wheels==3.30.1.post3
async-timeout==3.0.1
attrs==19.3.0
Babel==2.8.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==3.0.4
click==7.1.2
colorama==0.4.3
colour==0.1.5
commonmark==0.9.1
contextlib2==0.6.0.post1
discord.py==1.5.1
distro==1.5.0
fuzzywuzzy==0.18.0
h11==0.12.0
httpcore==0.12.2
httpx==0.16.1
humanize==3.2.0
idna==2.10
Markdown==3.2.2
motor==2.3.0
multidict==4.7.6
pip-autoremove==0.9.1
praw==7.1.0
prawcore==1.5.0
Pygments==2.6.1
pymongo==3.11.2
python-dateutil==2.8.1
python-Levenshtein-wheels==0.13.1
pytz==2020.1
PyYAML==5.3.1
rfc3986==1.4.0
rich==9.5.1
schema==0.7.2
six==1.15.0
sniffio==1.2.0
soupsieve==2.1
tqdm==4.48.0
typing-extensions==3.7.4.2
update-checker==0.18.0
urllib3==1.26.2
uvloop==0.14.0
websocket-client==0.57.0
yarl==1.5.1
